services/api-fastapi/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
psycopg[binary]>=3.2.1
pydantic>=2.10.0
redis>=5.0.8
openai>=1.40.0
python-dotenv>=1.0.0
pytrends>=4.9.2
